Refactor Message to MessageDTO and deprecate errorMessage
